# Conflicts:
#	code/modules/power/singularity/emitter.dm

+++ code/game/gamemodes/changeling/implements/powers/body.dm
@@ -69,10 +69,16 @@
 		to_chat(src, "<span class='warning'>We cannot perform this ability in this form!</span>")
 		return
 
+	if(!isturf(loc)) // so people can't transform inside places they should not, like sleepers
+		return
+
 	if(H.handcuffed)
 		var/cuffs = H.handcuffed
 		H.u_equip(H.handcuffed)
 		qdel(cuffs)
+
+	if(H.buckled)
+		H.buckled.unbuckle_mob()
 
 	changeling.chem_charges--
 	H.visible_message("<span class='warning'>[H] transforms!</span>")
@@ -281,8 +287,8 @@
 	C.digitalcamo = !C.digitalcamo
 
 	spawn(0)
-		while(C && C.digitalcamo && C.mind && C.mind.changeling)
-			C.mind.changeling.chem_charges = max(C.mind.changeling.chem_charges - 1, 0)
+		while(C && C.digitalcamo && C.mind && changeling)
+			changeling.chem_charges = max(changeling.chem_charges - 1, 0)
 			sleep(40)
 
 	src.verbs -= /mob/proc/changeling_digitalcamo
@@ -299,7 +305,7 @@
 	var/datum/changeling/changeling = changeling_power(30, 0, 100, UNCONSCIOUS)
 	if(!changeling)
 		return FALSE
-	src.mind.changeling.chem_charges -= 30
+	changeling.chem_charges -= 30
 
 	var/mob/living/carbon/human/C = src
 	spawn(0)
@@ -360,11 +366,11 @@
 	feedback_add_details("changeling_powers","MV")
 
 	spawn(0)
-		while(src && src.mind && src.mind.changeling && src.mind.changeling.mimicing)
-			src.mind.changeling.chem_charges = max(src.mind.changeling.chem_charges - 1, 0)
+		while(src?.mind && changeling?.mimicing)
+			changeling.chem_charges = max(changeling.chem_charges - 1, 0)
 			sleep(40)
-		if(src && src.mind && src.mind.changeling)
-			src.mind.changeling.mimicing = ""
+		if(changeling)
+			changeling.mimicing = ""
 
 /mob/proc/armblades()
 	set category = "Changeling"
@@ -374,7 +380,7 @@
 	var/datum/changeling/changeling = changeling_power(20, 0, 0)
 	if(!changeling)
 		return FALSE
-	src.mind.changeling.chem_charges -= 20
+	changeling.chem_charges -= 20
 
 	var/mob/living/carbon/M = src
 
@@ -412,7 +418,7 @@
 	var/datum/changeling/changeling = changeling_power(20,0,0)
 	if(!changeling)
 		return FALSE
-	src.mind.changeling.chem_charges -= 20
+	changeling.chem_charges -= 20
 
 	var/mob/living/carbon/M = src
 
@@ -451,12 +457,15 @@
 	if(!changeling)
 		return FALSE
 
-	var/mob/living/M = src
+	if(!isturf(loc)) // so people can't transform inside places they should not, like sleepers
+		return
+
+	var/mob/living/carbon/human/M = src
 
 	if(alert("Are we sure we wish to reveal ourselves? This will only revert after ten minutes.", , "Yes", "No") == "No") //Changelings have to confirm whether they want to go full horrorform
 		return
 
-	src.mind.changeling.chem_charges -= 40
+	changeling.chem_charges -= 40
 
 	M.visible_message("<span class='danger'>[M] writhes and contorts, their body expanding to inhuman proportions!</span>", \
 						"<span class='danger'>We begin our transformation to our true form!</span>")
@@ -471,10 +480,18 @@
 	var/mob/living/simple_animal/hostile/true_changeling/ling = new (get_turf(M))
 
 	if(istype(M,/mob/living/carbon/human))
+		if(M.handcuffed)
+			var/cuffs = M.handcuffed
+			M.u_equip(M.handcuffed)
+			qdel(cuffs)
+
 		for(var/obj/item/I in M.contents)
 			if(isorgan(I))
 				continue
 			M.drop_from_inventory(I)
+
+	if(M.buckled)
+		M.buckled.unbuckle_mob()
 
 	if(M.mind)
 		M.mind.transfer_to(ling)
